+++ src/molecules/ChipLarge.tsx
@@ -1,8 +1,17 @@
 import React from 'react';
-import { StyleSheet, Text, View, ViewStyle } from 'react-native';
+import {
+  StyleProp,
+  Image,
+  ImageRequireSource,
+  StyleSheet,
+  Text,
+  View,
+  ViewStyle,
+} from 'react-native';
 import { MaybeTouchable } from 'src/components/MaybeTouchable';
 import { colors, isLight } from 'src/utils/colors';
 import { Icon } from 'src/utils/const';
+import { layoutStyles } from 'src/utils/styles';
 import { registeredTextStyle } from 'src/utils/textStyles';
 
 const styles = StyleSheet.create({
@@ -13,19 +22,22 @@ const styles = StyleSheet.create({
     height: 22,
   } as ViewStyle,
   IconAndText: {
-    flexDirection: 'row',
-    alignItems: 'center',
+    ...layoutStyles.RowAlignCenter,
   },
   Icon: {
     marginRight: 4,
   },
   Text: registeredTextStyle('titleBold', colors.ulisse, 'chip-text'),
   TextDark: registeredTextStyle('titleBold', colors.uma, 'chip-text-dark'),
+  topLeftMargins: {
+    marginRight: 2,
+    marginTop: 4,
+  },
 });
 
 export type ChipLargeProps = {
   label: string;
-  icon?: string;
+  icon?: ImageRequireSource | string;
   color: string;
   /**
    * If specified, forces text to be dark (can be used to skip color lightness computation
@@ -33,18 +45,38 @@ export type ChipLargeProps = {
    */
   colorIsLight?: boolean;
   onPress?: () => any;
+  containerStyle?: 'topRightMargins' | 'topLeftMargins';
 };
 
-const renderChip = (color: string, darkText: boolean, label: string, icon?: string) => (
-  <View style={[styles.Wrapper, { backgroundColor: color }]}>
+const renderImageOrIcon = (
+  image: ImageRequireSource | string,
+  darkText: boolean,
+  style: StyleProp<ViewStyle>
+) => {
+  if (image === 'walk') return <Icon name={image} size={20} color={colors.uto} />;
+  return typeof image === 'string' ? (
+    <Icon name={image} size={20} color={darkText ? colors.uma : colors.ulisse} style={style} />
+  ) : (
+    <Image resizeMethod="scale" source={image} style={{ width: 22, height: 22 }} />
+  );
+};
+
+const renderChip = (
+  color: string,
+  darkText: boolean,
+  label: string,
+  icon?: ImageRequireSource | string,
+  containerStyle?: ChipLargeProps['containerStyle']
+) => (
+  <View
+    style={[
+      { ...styles.Wrapper, backgroundColor: color },
+      containerStyle && { ...styles[containerStyle] },
+    ]}
+  >
     {icon ? (
       <View style={styles.IconAndText}>
-        <Icon
-          style={styles.Icon}
-          name={icon}
-          size={20}
-          color={darkText ? colors.uma : colors.ulisse}
-        />
+        {renderImageOrIcon(icon, darkText, styles.Icon)}
         <Text style={darkText ? styles.TextDark : styles.Text}>{label.toUpperCase()}</Text>
       </View>
     ) : (
@@ -54,7 +86,7 @@ const renderChip = (color: string, darkText: boolean, label: string, icon?: stri
 );
 
 const ChipLargeUnmemoized = (props: ChipLargeProps) => {
-  const { color, colorIsLight, icon, label, onPress } = props;
+  const { color, colorIsLight, icon, label, onPress, containerStyle } = props;
   const darkText = colorIsLight ?? isLight(color);
   return onPress ? (
     <View style={{ flex: 1, alignItems: 'center' }}>
@@ -63,7 +95,7 @@ const ChipLargeUnmemoized = (props: ChipLargeProps) => {
       </MaybeTouchable>
     </View>
   ) : (
-    renderChip(color, darkText, label, icon)
+    renderChip(color, darkText, label, icon, containerStyle)
   );
 };