+++ src/utils/styles.ts
@@ -1,0 +1,13 @@
+import { StyleSheet } from 'react-native';
+export const layoutStyles = StyleSheet.create({
+  ColumnJustifyStart: {
+    display: 'flex',
+    flexDirection: 'column',
+    justifyContent: 'flex-start',
+  },
+  RowAlignCenter: {
+    display: 'flex',
+    flexDirection: 'row',
+    alignItems: 'center',
+  },
+});